stats: resize backing region for VNIC stats
The script that generates the statistics structures historically
doubled the size of the VNIC stats to accommodate for a byte
counter and packet count for each statistic provisioned in the
definition file. At some point the logic of this script was
reworked to maintain an absolute count, so doubling the size
of the VNIC stats structure is no longer required. This issue
was discovered by inspection while investigating wip://NIC-679.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>