cfg_act: fix preprocessor guards for missing PCIe islands
Functions cfg_act_bar_ptr() and cfg_act_vf_cfg_ptr() could reference
symbols that don't exist if the PCIe parameter was non-constant. The
first of these has an NFD provided equivalent in nfd_cfg_bar_base()
which works correctly, so this implementation is used instead. The
latter is updated to exclude the code for cases  where the compiler
unable to elide implementation due to the lack of constant arguments.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Signed-off-by: Edwin Peer <edwin.peer@netronome.com>