bpf: add flavor for BPF build
Introduce NS_FLAVOR_BPF and update NFD_NET_APP_ID to 2 for
BPF builds. In order not to clash with the BPF branch, the
new flavor has been omitted from the global FLAVOR_LIST
such that the BPF flavor is not automatically built by
the buildbot until such time as it is ready to replace
the existing BPF branch.

Signed-off-by: Mary Pham <mary.pham@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>