app_master: Update link state arrays for multi PCIE
Similar to a previous patch that expanded nic_control_word to support
multiple PCIe's, the arrays are changed to matrices with NFD_MAX_ISL
as the second dimension. In the case of "vs_lsc_list" this adds a third
dimension.

Arrays changed:
vs_current
ls_current
pending
vf_lsc_list

At this stage, the constant NIC_PCI is used as the PCIe index into the
matrices. This will become variable in future patches.

Change-Id: I2f5934fbc612e9bc8f58ce00a751ae019cac17b5
Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Frederik Lotter <frederik.lotter@netronome.com>
Reviewed-by: Pablo Cascón <pablo.cascon@netronome.com>
Reviewed-by: Michael Rapson <michael.rapson@netronome.com>