build: Added FLOORPLAN variable
This variable is specific to firmware/apps/nic/Makefile, and it is used
to determine the layout of the ME's.
E.g. All NFP-4xxx based CX NIC products share the same floorplan, but
NFP-6xxx LX NIC products can leverage extra ME islands and another PCIe,
and therefore have a different floorplan.

The SKU is chosen to determine the floorplan because it is the biggest
factor in determining the layout of the ME's.

Change-Id: I0f1483fe70e25dbba2f560ef89932d4be820debc
Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Michael Rapson <michael.rapson@netronome.com>
Reviewed-on: https://ci.netronome.com/r/1466
Tested-by: Zuul CI <zuul@ci.netronome.com>