actions: Add the L2 switch (wire) action.
This action is used on multi-host platforms. It makes the initial
determination about where a packet that ingresses on the NBI should go.

It uses the MU Lookup engine to do a 48 bit lookup on the packet
destination MAC address. If the DST MAC does not match any entry in the
table, the packet is dropped.

Tests added:
DST MAC match
DST MAC no match
Broadcast/Multicast

Change-Id: I545f3fe4ace1505ea48b048e56224df1cf3f205f
Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Michael Rapson <michael.rapson@netronome.com>
Reviewed-on: https://ci.netronome.com/r/1558
Tested-by: Zuul CI <zuul@ci.netronome.com>