actions: Add the L2 switch (host) action.
This action is used on multi host platforms. It makes the initial
determination about where a packet that ingresses on NFD should
go.

It uses the MU Lookup engine to do a 48 bit lookup on the packet
destination MAC address. If the DST MAC does not match any entry in
the table, the action processing continues as normal

Tests added:
DST MAC match
DST MAC no match
Broadcast/Multicast

Change-Id: Ic1921755182ed05cbddc29804b3328b5f14b5a5e
Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Michael Rapson <michael.rapson@netronome.com>
Reviewed-on: https://ci.netronome.com/r/1559
Tested-by: Zuul CI <zuul@ci.netronome.com>