nfd: add patch to always free CTM buffer in PCI_OUT
The NFD_OUT_ALWAYS_HAS_CTM feature that made it into upstream NFD went
one step further than what CoreNIC desires by extending the assumptions
about the existence of a CTM buffer to the DMA logic. The cause of this
may have been the inappropriate naming of the feature to begin with.

CoreNIC exploits the NFP split buffer architecture in order to feign
two versions of a packet for multicast. To send the one version to the
host, CoreNIC tells NFD that there is indeed a CTM buffer. To send the
other version of the packet (say with a different VLAN tag in the head
part of the packet), NFD is told that there is no CTM buffer (expecting
an MU only DMA to occur). In both cases, there is still a CTM buffer
allocated, even if it is not used as a DMA source, which is expected to
be freed by NFD (except when BLQ == 3). When sending the MU only version
of the packet, CoreNIC essentially lies to NFD about the CTM buffer's
existence.

Introducing an NFD_OUT_ALWAYS_FREE_CTM optimization (accordingly named)
to restrict the assumption of the existence of a CTM buffer to the DMA
completion logic better reflects CoreNIC's intended behavior.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>