hashmap: reduce macro arguments in control message handler
The cmsg_proc() macro's argument list has become long and unwieldly
with plans to add further parameterized capabilities later. Since
this is a reasonably top level macro, some of the arguments are
promoted to the calling scope with consequent renames to make the
more global nature of these registers clear. The balance of the
code is refactored to take this primary change into account.

Signed-off-by: Mary Pham <mary.pham@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>