hashmap: use atomics to access TID field
The table ID field shares a cache line with other data elements
that are accessed using the atomic engine. When adding new entry,
use atomic_write instead of write32. Also, the TID field should
be cleared when invalidating an entry. This can be accomplished
by widening the existing atomic subtract which is performed to
release the entry lock.

Signed-off-by: Mary Pham <mary.pham@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>