[build] move to latest SDK
Newer versions of the SDK use nfcc-ng by default, breaking the build
when using the latest toolchain. Absent rework, app master now requires
the "-v1" nfcc option in order to force the use of the legacy compiler.
Also update the build to depend on a later toolchain since only newer
versions of the SDK support this added compatibility argument.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>