rss: remove unnecessary NOPs when restoring T_INDEX
Fill the wasted NOPs with useful work. This change is still a net win
for the more rarely used RSSv1 metadata protocol even though only one
of the branch defer slots is now gainfully filled.

The branch to the end# label for packets having unknown L4 was also
incorrect. Not only should RSS still be performed on the valid L3
portion of the header in such cases, but branching to the end of the
macro would result in T_INDEX not being properly restored over
potential context switches.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>