rss: exploit CLS read to create code hole
pv_seek() consumes 10 cycles in the fast path when used to continue
execution at a requested label. In contrast, the non-branching
instantiation requires 8 cycles. Thus, two cycles are saved when
performing RSS on TCP packets (which is typical) in exchange for
the additional cycle in the common path. This change also benefits
UDP packets when RSS is configured accordingly (not the default).
In the case where RSS is only performed over L3, the action has
less work to do overall and so the extra cycle seems a reasonable
tradeoff.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>