rss: fix context swap bug that could corrupt CRC
pv_seek() might swap when accessing the L4 header. In this case, another
thread may reset global CRC state corrupting the partially calculated
value.

The issue is addressed by refactoring the code to preread the L4 header
and cache it so that the CRC calculation can be performed later without
switching context. Saving and restoring CRC context was also considered,
however, this would be more costly in terms of access latencies to the
CRC unit.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>