test: Do not build *.obj's for microC unit testing
The problem with the prior approach was that the *.obj's could be build
with different constant values than the actual unit test. NFCC does not
check for this. Consequently there could be subtle bugs in the tests.

Each of the tests must #include the .c source file(s) with the functions it
is testing.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>