nic_tables: fix min RX buffer size in VLAN member for small MRUs
We previously only updated the lowest receive buffer size for all VFs having a
common VLAN if current minimum buffer size was non zero. However a zero value
is a valid size for MRUs up to 256B and does not necessarily mean there are no
VFs currently up. We were allowing a VF with MRU larger than 256B to override
the minimum receive buffer of a VF with an MRU below 256B. The firmware would
then allow packets larger than the available size through to the host causing
a system crash.

resolves://NIC-789

Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>