scripts: Add options to soft reset and/or link BLM during tests.
Some future unit tests will require that the BLM ME be loaded during the
test. This provides an option for the individual tests to declare
"TEST_REQ_BLM" in the comments. BLM will be assembled and linked with
the test firmware files.

Some future unit tests try to allocate CTM buffers. If the chip is not
soft reset, the CTM might be in a state where all the buffers are
allocated, and the test will fail. This option can be enabled by adding
"TEST_REQ_RESET" in the comments.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Mary Pham <mary.pham@netronome.com>