app_config_tables: remove duplicate include directives
Keeping #include <app_mac_vlan_config_cmsg.h> in app_config_tables.h
because the latter contains a function prototype that references a
struct that is declared in the former. That way the warning will not
resurface if the include order in app_config_tables.c is changed.

Fixed function return type of cfg_act_write_veb so that the prototype
matches the function implementation.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>