app_master: fix clearing top of VF MAC address on vf_cfg symbol
When the PF sets the VF MAC address almost everything works with the
VF cfg BAR getting the right/complete MAC address. However writing the
return error code on the vf_cfg symbol clears the top 2 bytes of the
MAC address. That makes the driver read incomplete VF's MAC addresses
and "ip link" to display that. This was detected while running driver
tests.unit.sriov_ndos

Fix the bug by writing the return error code to the right
address/offset.

Additionally move the read of the VF's MAC address to only be done if
it is actually needed

resolves://NIC-816

Fixes: 032567a811c5 ("build: update to latest NFD, dropping local patches.")
Signed-off-by: Pablo Cascón <pablo.cascon@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>
Acked-by: Edwin Peer <edwin.peer@netronome.com>