build: add sodium card support
Add build target support for both variants of the sodium card (with
and without the crypto island).

We are getting more and more customer requests for eBPF and CoreNIC
FX (Sodium) card evaluations. It makes sense to have this target
available in CoreNIC, since basic functionality from the ARM side
(PCIe0) works.

Note that the CoreNIC firmware host side (PCIe1) support is not yet
available today.

Signed-off-by: Fred Lotter <frederik.lotter@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>