config: Set CTM used for packets to 1/2 of CTM and rebalance credits.
If CTM is used for symbols and packets, the
xpb:CTMXpbMap.MuPacketReg.MUPEMemConfig register must be set to 1 per
CTM island to indicate this to the packet engine. Without this setting,
there is no guarantee that the PE's will not allocate packets that are
in the area reserved for symbols. The setting of 1 indicates 1/2 CTM is
used for packets. Other options are 1/4 and 1/1.

1/2 CTM is 128K, this translates into 64 2K buffers. Since the NBI DMA
buffer credit count is allowed to go -1 for one cycle, there are 63 buffer
credits to divide between NBI and ME's.

In the TX direction, the ME's allocate the CTM buffers. 12 ME's per
island, running in 4 context mode require 48 packet credits. The buffer
size in this case is always 256B. A minimum of 6 buffer credits are
required (48*256/2048=6).

The allocation logic (both in hardware and software) allocates a full 2K
buffer, even for a 256B request. If the 2K buffer is not used in its
entirety, the credit is returned, BUT only on the next allocation request.
To prevent unnecessary delays due to retrying for buffer credits, more than
6 are assigned at startup. 12 ME threads on an island can send allocation
requests simultaneously, thus a minimum of 12 buffer credits are
assigned.

The surplus buffer credits are assigned to the NBI DMA BPE.
The final value is 63-12=51.

Note 1: Attempting to over-allocate CTM buffers will cause allocation
requests to stall waiting for buffers to be freed. This can lead to
deadlocks between GRO and NBI DMA as seen in comment://NIC-588

Note 2: Some of the current control message code allocates CTM buffers
This is not needed because these packets are not destined to the TM.

resolves://NIC-812
resolves://NIC-721

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>