app_master: add VF trusted support
Advertise the new NFD capability. Allow a trusted VF to change its MAC
address even after being administratively set by the PF. Add the new
MAC to the VEB if the change is accepted. Return an error otherwise.

Additionally, allow a non trusted VF to change its MAC if the PF has not
set it yet (all 0s), useful for lab testing.

There is a known corner case when trusted mode is reset back to off
after the VF changed its MAC address and the link is admin set to down
and up again. It is deemed not relevant for production or lab testing.

There is a known corner case when trusted mode is set back to off
after the VF changed its MAC address and its link is set admin down
and up. In this case a "reconfig error" will appear on the host's
message log.

Signed-off-by: Pablo Cascón <pablo.cascon@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>