docs: refine the process for obtaining the Linux toolchain
To assist the support team, the emails to the help@netronome.com address should
differentiate between requests for the full Windows SDK and the Linux toolchain
required to build CoreNIC. RST doesn't appear to provide a way to natively
handle parameterization of mailto links, thus we resort to raw HTML for this.

The user does not have to review and accept the license terms before the tools
are downloaded, provided that the following notice is included when we provide
access:

BY LOADING OR USING THE SOFTWARE, YOU AS AN INDIVIDUAL AND ON BEHALF OF YOUR
EMPLOYER ("YOU") AGREE TO THE TERMS OF THIS AGREEMENT. IF YOU DO NOT WISH TO
SO AGREE, DO NOT INSTALL, OR IF INSTALLED, DELETE ALL COPIES OF THE SOFTWARE
AND DO NOT USE THE SOFTWARE.

This simplifies the process to obtain the toolchain while ensuring that the
license terms are always consistent with those included in the toolchain
package. A user who does not wish to agree must simply delete the files and
refrain from using them.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>