tests: Update the VF reconfig unit tests for trusted VF support
Since the trusted VF support, extra setup is needed per VF in order to
run the tests. Notably the VF's MAC must be in the BAR before reconfig
(as it would in the real world, because either the PF set is via
firmware or the VF would set it).

Separate unit tests for the trusted VF feature are planned for the
future.

Some TODO comments and whitespace fixes are included.

Signed-off-by: Johan Moraal <johan.moraal@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>
Reviewed-by: Pablo Cascón <pablo.cascon@netronome.com>