[make] allow SDK detection and override
Add an explicit check to see if the SDK binaries can be found.
If not, present the user with a hint on how to resolve the issue.

"SDK not found. Please update the SDK path variable:
export NETRONOME=<path>"

Add support to override the SDK version check. This will simplify
development and testing with new versions of the SDK.

export REQ_SDK_VER=<version>
export REQ_SDK_BUILD=<build>

Fix the firmware_show_tests target.

Note that if Netronome updates the required SDK version, then it
is after Netronome tested that version against the nic-firmware
project. Netronome cannot guarantee operation of firmware that was
not build against the offical SDK version specified in the Makefile.

Change-Id: Ic81cd34bc8244c470e71cd2b421f9f89399370dd
Signed-off-by: Fred Lotter <frederik.lotter@netronome.com>
Reviewed-by: Johan Moraal <johan.moraal@netronome.com>