Decrease workers and move BLM for more pkt bufs
Packet buffers per island increased from 24 to 32, and workers decreased
to 8 per island. This frees up 3 MEs, and it means that each CTM will
have enough buffers for all workers to be serviced. Although the
decrease in workers seems suboptimal, at this stage it seems unavoidable
for a functional system that doesn't have the potential to get starved
of buffers.

For nfp-6000/4000, assign ila0.me0 to blm, same as before