app_config_tables: Move veb_stored_keys out of CTM.
Restrictions placed on symbols destined for worker island CTM on NFP3800
platforms makes it impossible to reserve more than 64K for packet
buffers: packet buffers need to have an offset of zero (ctm+0), and the
32K vf_vlan_cache symbol needs to be 64K aligned. This means that 32K
CTM memory is available for veb_stored_keys (24B) or
nic_vlan_to_vnic_map_tbl (32KB) (clearly both won't fit). If the former
is moved to emem, the increased I/O latency only comes into play when a
VF is upped. This is preferred over moving the latter to emem, as I/O
to/from this table will affect the datapath.