test: make unit test support chip nfp-38xxc
The original run_tests.sh script is only support NFP4000/6000.
To support NFP3800, we need to get the targat board type from
nfp-hwinfo, and compile and link with the specific chip type.
And the assignment of blm function block is moved to pcie0.me7
if it is NFP3800, since there is no island 48.