app_master: Fix pointer truncation for NFCC V2.
When __mem is used instead of __emem in an attempt to keep code
compatible with all chips (without the use of unnecessary preproc
tests), and NFCC V2 is used (as is the case for Kestrel chips), the
default address length is 32b, potentially truncating 40b emem
addresses. __mem40 is used instead to explicitly declare the symbol or
pointer as a 40b address.