actions: rearrange action instruction table in cls
Half of the cls memory in NFP6XXX is allocated to action instruction
table, that is 32K. However the total size of cls in NFP38XX is 32K,
we need rearrange the cls memory here.

Since there's only one pcie and maximum 128 queues in NFP38XX, 128*64=8K
is allcoated to NFD part. And with limited port number, 16*64=1K is
enough for NBI part.  Considering that action instruction table address
is used to determine where the packet is from, NFD or NBI, we'd better
keep the base offset of NFD(0) and NBI(16K) as it is.

Originally the adjacent cls space next to instruction table is allocated
to "NIC_RSS_TBL_ADDR", whose address need be power of 2, so it is 32K,
which exceeds cls size. Now we can place it into the gap inbetween action
instruction table of NFD and NBI.

Here's the new scheme:
    0   - 8K   action instruction table for NFD
    8K  - ?    rss table, currently allocated 256 bytes
    16K - 17K  action instruction table for NBI
This scheme is for 3800 only, no change for other platforms.

Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>