pv: modify pm and gro discriptor for kestrel
In NFP38XX, the pm is disabled by default, so we shouldn't prepend
any pms to packet data, only 4 bytes mac prepend is remained. Also
tm_pm_init.list is not needed in NFP38xx, remove it in Makefile.

And the definition of packet ready command is different in NFP38XX
from that in NFP6XXX according to the databook, so we need modify
the gro discriptor accordingly, which is used by nbi instructions
to send packets out.

Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>