EMU: Enable bulk engine byte mask swaping
Accoding to databoot chapter 8.2.2.3.2.1, the nfp-3800 introduces
a mode that provides the ability to swap the byte mask along with
the data when byte swapping is used. And this function applies
only to commands where byte swap is used, and only when length
field is 0.
I think this explains why the packets are malformed when the it
is not 4 bytes aligned. Because in this case, the byte mask is
used to indicate which bytes are to be written, and if the data
is byte swaped, while byte mask is not swaped, bulk engine will
chose the wrong bytes.